// FE/app/lib/types/JobInfo.ts
export interface JobInfo {
    id: string;
    title: string;
    company: string;
    company_logo: string;
    address: string;
    career: string;
    level: string;
    form: string;
    salary: string;
    experience: string;
    user_upload: string;
    avatar: string;
    quantity: number;
    apply_from: string;
    apply_to: string;
  }